+++ gym_trading/tests/test_meta_broker.py
@@ -18,8 +18,8 @@ class MetaBrokerTestCases(unittest.TestCase):
         )
 
     def random_init(self):
-        currency_amounts = np.random.choice(10000,len(self.broker.portfolio.currencies)) / np.float64(10)
-        exchange_rates = np.random.choice(10000,len(self.broker.portfolio.exchanges)) / np.float64(10.0)
+        currency_amounts = np.random.choice(10000,len(self.broker.portfolio.currencies)) / np.float32(10)
+        exchange_rates = np.random.choice(10000,len(self.broker.portfolio.exchanges)) / np.float32(10.0)
         self.final_init(currency_amounts, exchange_rates)
 
     def zero_init(self):
@@ -30,7 +30,7 @@ class MetaBrokerTestCases(unittest.TestCase):
     def final_init(self, currency_amounts, exchange_rates):
         self.inventory = dict(zip(self.broker.portfolio.currencies, currency_amounts))
         self.bid_ask_prices = dict(zip(self.broker.portfolio.exchanges,
-                                  [{'ask': np.float64(v) * np.float64(1.00001), 'bid': np.float64(v) * np.float64(0.99999)} for v in exchange_rates]))
+                                  [{'ask': np.float32(v) * np.float32(1.00001), 'bid': np.float32(v) * np.float32(0.99999)} for v in exchange_rates]))
         self.broker.initialize(self.bid_ask_prices, self.inventory)
 
     def test_step_updates_portfolio_attributes(self):
@@ -38,7 +38,7 @@ class MetaBrokerTestCases(unittest.TestCase):
         total_value_before_step = self.broker.portfolio.total_value
         pnl_before_step = self.broker.portfolio.pnl
         for i in range(5):
-            vals2 = np.random.choice(10000,len(self.broker.portfolio.currencies)) / np.float64(10.0)
+            vals2 = np.random.choice(10000,len(self.broker.portfolio.currencies)) / np.float32(10.0)
             bid_prices = dict(zip(self.broker.portfolio.currencies, vals2))
             self.broker.portfolio.step(bid_prices)
         total_value_after_step = self.broker.portfolio.total_value
@@ -49,9 +49,9 @@ class MetaBrokerTestCases(unittest.TestCase):
     def test_add_order_updates_portfolio_attributes(self):
         self.zero_init()
         bp = deepcopy(self.bid_ask_prices)
-        bp['BTC-USD'] = {'bid': np.float64(11000.0), 'ask': np.float64(11000.01)}
+        bp['BTC-USD'] = {'bid': np.float32(11000.0), 'ask': np.float32(11000.01)}
         inv = deepcopy(self.inventory)
-        inv['USD'] = np.float64(11000.0)
+        inv['USD'] = np.float32(11000.0)
         self.broker.initialize(bp, inv)
 
         total_value_before_order = self.broker.portfolio.total_value
@@ -59,7 +59,7 @@ class MetaBrokerTestCases(unittest.TestCase):
         realized_value_before_order = self.broker.portfolio.realized_value
         realized_pnl_before_order = self.broker.portfolio.realized_pnl
         
-        order = MarketOrder(ccy='BTC-USD', side='long', price=np.float64(11000.0), size=np.float64(1.0))
+        order = MarketOrder(ccy='BTC-USD', side='long', price=np.float32(11000.0), size=np.float32(1.0))
 
         self.assertTrue(self.broker.portfolio.add_order(order))
 
@@ -75,32 +75,32 @@ class MetaBrokerTestCases(unittest.TestCase):
 
     def test_reallocation(self):
         exchange_rates = [
-            np.float64(11812.31),
-            np.float64(411.30),
-            np.float64(15.22256),
-            np.float64(0.105917),
-            np.float64(0.03478),
-            np.float64(0.00000898),
-            np.float64(0.03709380),
+            np.float32(11812.31),
+            np.float32(411.30),
+            np.float32(15.22256),
+            np.float32(0.105917),
+            np.float32(0.03478),
+            np.float32(0.00000898),
+            np.float32(0.03709380),
         ]
         bid_ask_prices = dict(zip(EXCHANGES,
-                                  [{'ask': v * np.float64(1.00001), 'bid': v * np.float64(0.99999)} for v in exchange_rates]))
+                                  [{'ask': v * np.float32(1.00001), 'bid': v * np.float32(0.99999)} for v in exchange_rates]))
         inventory = {}
         for c in self.broker.portfolio.currencies:
             if c == 'USD':
-                inventory[c] = np.float64(10000.0)
+                inventory[c] = np.float32(10000.0)
             else:
-                inventory[c] = np.float64(10000.0) / bid_ask_prices[f'{c}-USD']['bid']
+                inventory[c] = np.float32(10000.0) / bid_ask_prices[f'{c}-USD']['bid']
 
         self.broker.initialize(bid_ask_prices, inventory)
 
-        cts = {True: np.float64(0.0), False: np.float64(0.0)}
+        cts = {True: np.float32(0.0), False: np.float32(0.0)}
         for _ in range(1000):
             initial_allocation = self.broker.allocation
 
             currency_range = range(len(self.broker.portfolio.currencies))
             
-            nums = np.array([np.float64(np.random.uniform()) for _ in currency_range])
+            nums = np.array([np.float32(np.random.uniform()) for _ in currency_range])
             total = sum(nums)
 
             prior_allocation = self.broker.allocation
@@ -119,26 +119,26 @@ class MetaBrokerTestCases(unittest.TestCase):
                     "DIFF": allocation_diffs,
                     "TRADES": self.broker.portfolio.total_trade_count - prior_trades,
                 }
-                self.assertLessEqual(max(allocation_diffs), np.float64(0.01), f"should never miss target by more than 1%.\n{pformat(logs)}")
-            cts[reached_target] += np.float64(1.0)
+                self.assertLessEqual(max(allocation_diffs), np.float32(0.01), f"should never miss target by more than 1%.\n{pformat(logs)}")
+            cts[reached_target] += np.float32(1.0)
 
         success_rate = cts[True] / (cts[True] + cts[False])
         #print("SUCCESS_RATE", success_rate)
-        self.assertGreaterEqual(success_rate, np.float64(0.75), 'should hit target more than 75% of the time')
+        self.assertGreaterEqual(success_rate, np.float32(0.75), 'should hit target more than 75% of the time')
 
     def test_get_statistics(self):
         self.zero_init()
         bp = deepcopy(self.bid_ask_prices)
-        bp['BTC-USD'] = {'bid': np.float64(10000.0), 'ask': np.float64(10000.01)}
+        bp['BTC-USD'] = {'bid': np.float32(10000.0), 'ask': np.float32(10000.01)}
         inv = deepcopy(self.inventory)
-        inv['USD'] = np.float64(20000.0)
+        inv['USD'] = np.float32(20000.0)
         self.broker.initialize(bp, inv)
         
-        order = MarketOrder(ccy='BTC-USD', side='long', price=np.float64(10000.0), size=np.float64(1.0))
+        order = MarketOrder(ccy='BTC-USD', side='long', price=np.float32(10000.0), size=np.float32(1.0))
 
         self.assertTrue(self.broker.portfolio.add_order(order))
 
-        bp['BTC-USD'] = {'bid': np.float64(10100.0), 'ask': np.float64(10100.01)}
+        bp['BTC-USD'] = {'bid': np.float32(10100.0), 'ask': np.float32(10100.01)}
         self.broker.step(bp)
 
         expected_stats = {
