+++ gym_trading/utils/portfolio.py
@@ -19,15 +19,15 @@ class Portfolio(object):
                  cryptos: List[str] = ['ETH', 'BTC'],
                  exchanges: List[str] = ['BTC-USD', 'ETH-USD', 'ETH-BTC'],
                  transaction_fee: bool = True,
-                 initial_inventory: Optional[Dict[str, np.float64]] = None,
-                 initial_bid_prices: Optional[Dict[str, np.float64]] = None):
+                 initial_inventory: Optional[Dict[str, np.float32]] = None,
+                 initial_bid_prices: Optional[Dict[str, np.float32]] = None):
         """
         :param fiat: (str)
         :param cryptos: (List[str])
         :param exchanges: (List[str])
         :param transaction_fee: (bool)
-        :param initial_inventory: (Optional[Dict[str, np.float64]])
-        :param initial_bid_prices: (Optional[Dict[str, np.float64]])
+        :param initial_inventory: (Optional[Dict[str, np.float32]])
+        :param initial_bid_prices: (Optional[Dict[str, np.float32]])
         """
         self.transaction_fee = transaction_fee
         self.fiat = fiat
@@ -50,21 +50,21 @@ class Portfolio(object):
 
         :return: (void)
         """
-        self.initial_total_value = np.float64(0)
-        self.initial_realized_value = np.float64(0)
+        self.initial_total_value = np.float32(0)
+        self.initial_realized_value = np.float32(0)
         self.trades.clear()
         self.statistics.reset()
-        self.inventory = {c: np.float64(0) for c in self.currencies}
-        self.bid_prices = {c: np.float64(0) for c in self.cryptos}
-        self.bid_prices[self.fiat] = np.float64(1)
+        self.inventory = {c: np.float32(0) for c in self.currencies}
+        self.bid_prices = {c: np.float32(0) for c in self.cryptos}
+        self.bid_prices[self.fiat] = np.float32(1)
 
-    def initialize(self, bid_prices: Dict[str, np.float64], inventory: Dict[str, np.float64] = {}) -> None:
+    def initialize(self, bid_prices: Dict[str, np.float32], inventory: Dict[str, np.float32] = {}) -> None:
         """
         Adds starting values to the portfolio.
 
-        :param inventory: (Dict[str, np.float64]) maps currenies to starting amount
-        :param bid_prices: (Dict[str, np.float64]) the most recent bid prices for each
-            crypto-fiat exchange. Keys are the crypto symbol, bid_prices[fiat] == np.float64(1)
+        :param inventory: (Dict[str, np.float32]) maps currenies to starting amount
+        :param bid_prices: (Dict[str, np.float32]) the most recent bid prices for each
+            crypto-fiat exchange. Keys are the crypto symbol, bid_prices[fiat] == np.float32(1)
         """
         self._initialize_bid_prices(bid_prices)
         self._initialize_inventory(inventory)
@@ -80,12 +80,12 @@ class Portfolio(object):
         return msg
 
     @property
-    def value_breakdown(self) -> Dict[str, np.float64]:
+    def value_breakdown(self) -> Dict[str, np.float32]:
         """
         Returns a dictionary of each position mapped to it's 
         current value in fiat currency.
 
-        :return: (Dict[str, np.float64]) portfolio value breakdown
+        :return: (Dict[str, np.float32]) portfolio value breakdown
         """
         return {
             currency: (count * self.bid_prices[currency])
@@ -93,68 +93,68 @@ class Portfolio(object):
         }
 
     @property
-    def unrealized_value(self) -> np.float64:
+    def unrealized_value(self) -> np.float32:
         """
         calculates the value of cryptos in portfolio by fiat currency
 
-        :return: (np.float64) portfolio value (e.g. USD)
+        :return: (np.float32) portfolio value (e.g. USD)
         """
         return sum([self.value_breakdown[c] for c in self.cryptos])
 
     @property
-    def realized_value(self) -> np.float64:
+    def realized_value(self) -> np.float32:
         """
-        :return: (np.float64) value of fiat holdings in portfolio
+        :return: (np.float32) value of fiat holdings in portfolio
         """
         return copy(self.inventory[self.fiat])
 
     @property
-    def total_value(self) -> np.float64:
+    def total_value(self) -> np.float32:
         """
         calculates the total value of portfolio in fiat currency
 
-        :return: (np.float64) portfolio value (e.g. USD)
+        :return: (np.float32) portfolio value (e.g. USD)
         """
         return sum(list(self.value_breakdown.values()))
 
     @property
-    def allocation(self) -> Dict[str, np.float64]:
+    def allocation(self) -> Dict[str, np.float32]:
         """
         The fractional breakdown of currencies by fiat value
 
-        :return: (Dict[str, np.float64]) portfolio allocation - sum(allocation.values()) == np.float64(1)
+        :return: (Dict[str, np.float32]) portfolio allocation - sum(allocation.values()) == np.float32(1)
         """
-        if self.total_value <= np.float64(0):
-            return {c: np.float64(0) for c in self.currencies}
-        return {c: np.float64(v / self.total_value) for c, v in self.value_breakdown.items()}
+        if self.total_value <= np.float32(0):
+            return {c: np.float32(0) for c in self.currencies}
+        return {c: np.float32(v / self.total_value) for c, v in self.value_breakdown.items()}
 
     @property
     def total_trade_count(self) -> int:
         return self.trades.__len__()
 
     @property
-    def pnl(self) -> np.float64:
-        if self.initial_total_value != np.float64(0):
-            return (self.total_value / self.initial_total_value) - np.float64(1)
-        return np.float64(0)
+    def pnl(self) -> np.float32:
+        if self.initial_total_value != np.float32(0):
+            return (self.total_value / self.initial_total_value) - np.float32(1)
+        return np.float32(0)
 
     @property
-    def realized_pnl(self) -> np.float64:
-        if self.initial_realized_value != np.float64(0):
-            return (self.realized_value / self.initial_realized_value) - np.float64(1)
-        return np.float64(0)
+    def realized_pnl(self) -> np.float32:
+        if self.initial_realized_value != np.float32(0):
+            return (self.realized_value / self.initial_realized_value) - np.float32(1)
+        return np.float32(0)
 
-    def _validate_inventory(self, inventory: Dict[str, np.float64]) -> None:
+    def _validate_inventory(self, inventory: Dict[str, np.float32]) -> None:
         invalid_inventory_items = [k not in self.currencies for k in inventory]
         if any(invalid_inventory_items):
             raise ValueError("inventory contains unknown currency: " + \
                              f"{list(inventory.keys())[invalid_inventory_items.index(True)]}")
 
-    def _initialize_inventory(self, inventory: Dict[str, np.float64] = {}) -> None:
+    def _initialize_inventory(self, inventory: Dict[str, np.float32] = {}) -> None:
         self._validate_inventory(inventory)
         self.inventory.update(inventory)
 
-    def _validate_bid_prices(self, bid_prices: Dict[str, np.float64]) -> None:
+    def _validate_bid_prices(self, bid_prices: Dict[str, np.float32]) -> None:
         invalid_bid_prices = [k not in self.currencies for k in bid_prices]
         missing_bid_prices = [k not in bid_prices for k in self.currencies]
         if any(invalid_bid_prices):
@@ -164,12 +164,12 @@ class Portfolio(object):
             raise ValueError("Missing bid price for currency: " + \
                              f"{list(self.currencies.keys())[missing_bid_prices.index(True)]}")
 
-    def _initialize_bid_prices(self, bid_prices: Dict[str, np.float64]) -> None:
+    def _initialize_bid_prices(self, bid_prices: Dict[str, np.float32]) -> None:
         self._validate_bid_prices(bid_prices)
         self.bid_prices.update(bid_prices)
-        self.bid_prices[self.fiat] = np.float64(1)
+        self.bid_prices[self.fiat] = np.float32(1)
 
-    def step(self, bid_prices: Dict[str, np.float64]) -> None:
+    def step(self, bid_prices: Dict[str, np.float32]) -> None:
         """
         Step in environment and update portfolio value.
 
@@ -178,7 +178,7 @@ class Portfolio(object):
         """
         self._validate_bid_prices(bid_prices)
         self.bid_prices.update(bid_prices)
-        self.bid_prices[self.fiat] = np.float64(1)
+        self.bid_prices[self.fiat] = np.float32(1)
 
     def add_order(self, order: MarketOrder) -> bool:
         """
@@ -198,25 +198,25 @@ class Portfolio(object):
 
         # Create a hypothetical average execution price incorporating a fixed slippage
         if buying_asset:
-            order.average_execution_price = order.price * (np.float64(1) + SLIPPAGE)
+            order.average_execution_price = order.price * (np.float32(1) + SLIPPAGE)
         else:
-            order.average_execution_price = order.price * (np.float64(1) - SLIPPAGE)
+            order.average_execution_price = order.price * (np.float32(1) - SLIPPAGE)
 
         # Determine impact of order on inventory
         # ingress is where value is going
         ingress_sym = [asset, base][not int(buying_asset)]
         ingress_price = order.average_execution_price \
                      if not buying_asset \
-                     else np.float64(1)
+                     else np.float32(1)
         ingress_amount = order.size * ingress_price
         if self.transaction_fee:
-             ingress_amount *= np.float64(1) - MARKET_ORDER_FEE
+             ingress_amount *= np.float32(1) - MARKET_ORDER_FEE
 
         # egress is where value is coming from
         egress_sym = [asset, base][int(buying_asset)]
         egress_price = order.price \
                       if buying_asset \
-                      else np.float64(1)
+                      else np.float32(1)
         egress_amount = order.size * egress_price
 
         if self.inventory[egress_sym] < egress_amount:
